Save without erasing empty fields
I needed the possibility to save my template in a multiple times, due to separate pieces of code editing it. To do it I had to add the possibility to suspend the erasing of empty fields.

Don't know how to specify where to put the test files, nevertheless test is running fine.